refactor: p3at_plugin package sources and removing redundant algorithm files
