+++ p3at_plugin/CMakeLists.txt
@@ -1,8 +1,6 @@
 cmake_minimum_required(VERSION 2.8.3)
 project(p3at_plugin)
 
-# find_package(catkin REQUIRED COMPONENTS geometry_msgs tf std_msgs move_base_msgs actionlib)
-
 find_package(catkin REQUIRED COMPONENTS
   geometry_msgs
   tf 
@@ -14,14 +12,11 @@ find_package(catkin REQUIRED COMPONENTS
   geometry_msgs
   nav_msgs
 )
-
 find_package(PkgConfig REQUIRED)
 pkg_check_modules(TINYXML2 REQUIRED tinyxml2)
 
-# catkin_package( CATKIN_DEPENDS roscpp)
-
 catkin_package(
-  CATKIN_DEPENDS roscpp move_base_msgs actionlib geometry_msgs nav_msgs
+  CATKIN_DEPENDS roscpp move_base_msgs actionlib geometry_msgs nav_msgs 
 )
 
 ## Specify additional locations of header files
@@ -29,20 +24,14 @@ include_directories(
   include
   ${catkin_INCLUDE_DIRS}
   ${TINYXML2_INCLUDE_DIRS}
-  )
+)
 
 ## Declare a cpp executable (our new controller)
-add_executable(p3at src/p3at_controller_movebase.cc src/get_flower_locations.cc)
-add_executable(p3at_path_planning src/p3at_path_planning.cc src/get_flower_locations.cc)
-target_link_libraries(p3at ${catkin_LIBRARIES} tinyxml2)
-target_link_libraries(p3at_path_planning ${catkin_LIBRARIES} tinyxml2 ${roslib_LIBRARIES})
 
-## add_executable(actionlib src/actionlib_movebase.cc)
-## target_link_libraries(actionlib ${catkin_LIBRARIES})
+add_executable(waypoint_following_movebase src/waypoint_following_movebase.cc include/get_flower_locations.h include/waypoints.h)
+add_executable(waypoint_following_custom src/waypoint_following_custom.cc include/get_flower_locations.h include/waypoints.h)
+add_executable(path_planning_movebase src/path_planning_movebase.cc include/get_flower_locations.h)
 
-##add_executable(p3at src/p3at.cc)
-##target_link_libraries(p3at ${catkin_LIBRARIES})
-
-install(TARGETS p3at
-  RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
-)
+target_link_libraries(waypoint_following_movebase ${catkin_LIBRARIES} tinyxml2 ${roslib_LIBRARIES})
+target_link_libraries(waypoint_following_custom ${catkin_LIBRARIES} tinyxml2 ${roslib_LIBRARIES})
+target_link_libraries(path_planning_movebase ${catkin_LIBRARIES} tinyxml2 ${roslib_LIBRARIES})
