+++ p3at_plugin/package.xml
@@ -15,6 +15,7 @@
   <build_depend>actionlib</build_depend>
   <build_depend>nav_msgs</build_depend>
 
+
   <run_depend>urdf</run_depend>
   <run_depend>roscpp</run_depend>
   <run_depend>geometry_msgs</run_depend>
